requirements.txt
langchain==0.1.0
langchain-community>=0.0.10
python-dotenv>=0.19.0
beautifulsoup4>=4.11.1
requests>=2.28.1
scholarly>=1.7.11
pandas>=1.5.3
numpy>=1.23.5
tqdm>=4.65.0
flask>=2.0.1
flask-wtf>=1.0.0 